requirements.txt
annotated-types==0.7.0
anyio==3.7.1
attrs==25.3.0
backoff==2.2.1
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.2
colorama==0.4.6
distro==1.9.0
docstring_parser==0.17.0
filelock==3.18.0
filetype==1.2.0
fsspec==2025.7.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-vision==3.10.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
haystack-ai==2.16.1
haystack-experimental==0.12.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.3
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
lazy_imports==1.0.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
openai==1.99.1
packaging==25.0
pdf2image==1.17.0
pillow==11.3.0
posthog==6.4.0
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
PyMuPDF==1.26.3
pypdf==5.9.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
rfc3986==1.5.0
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.6.1
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.1.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.8.0
tqdm==4.67.1
transformers==4.55.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
streamlit
nltk>=3.9.1